When the list of '<items>' is empty hide listView.
The base listview that all the body items goes into
is a recycler sitting at the top of the view, even if
there's no items in it.  this means there's a (64dp?)
gap at the top of the screen where taps are ignored.

This fix hides the listview if the list size is zero,
and restores it if non-zero.

I have no idea what kind of test suite we can do here...but
I'd recommend it gets tested until it's silly.