WIP This is only a WIP and PoC...do *not* merge.
This build is meant to demonstrate the shake functionality. If you put a
breakpoint at JasonViewActivity.java:1467, launch with the script
indicated in the PR, you'll see that we do get to the $render callback,
but nothing seems to happen.

It is entirely possible that I've broken the script, or that I've broken
the code.

Thoughts?